move o useCse de adciona servico ao veiculo da entite veiculo para o veiculo-servico
