+++ src/modules/veiculos/infra/http/routers/veiculo.routers.ts
@@ -6,6 +6,5 @@ const router = Router();
 const veiculoController = VeiculoFactory.createVeiculoController();
 
 router.post('/create', (req, res) => veiculoController.create(req, res));
-router.post('/addServico', (req, res) => veiculoController.addServico(req, res));
 
 export default router;
