+++ src/modules/veiculos/infra/repositories/veiculoRepositorie.ts
@@ -1,7 +1,6 @@
 import { Veiculo, TiposVeiculos } from '@prisma/client';
 import prisma from '../../../../shared/infra/prisma/prisma';
 import { CreateVeiculoInterface } from '../../interfaces/createVeiculoInterface';
-import { AddServicosInterface } from '../../interfaces/addServicoInterface';
 
 interface VeiculoInput {
   placa: string;
@@ -9,7 +8,7 @@ interface VeiculoInput {
   user: number
 }
 
-export default class VeiculoRepository implements CreateVeiculoInterface, AddServicosInterface {
+export default class VeiculoRepository implements CreateVeiculoInterface {
   async create(data: VeiculoInput): Promise<Veiculo> {
     const veiculo = await prisma.veiculo.create({
       data: {
@@ -25,17 +24,5 @@ export default class VeiculoRepository implements CreateVeiculoInterface, AddSer
       tipo: veiculo.tipo,
       userId: veiculo.userId,
     };
-  }
-
-  async addServicos(veiculoId: number, servicoId: number): Promise<any> {
-    const agenda = await prisma.veiculoServico.create({
-      data: {
-        veiculoId,
-        servicoId,
-        dataInicio: new Date(),
-      },
-    });
-
-    return agenda;
   }
 }
